feat: move cards page and add home page
